Add network calls

+++ app/src/main/java/com/project/segunfrancis/bakingtime/model/Ingredient.java
@@ -8,7 +8,7 @@ import com.squareup.moshi.Json;
 public class Ingredient {
 
     @Json(name = "quantity")
-    private Integer quantity;
+    private double quantity;
 
     @Json(name = "measure")
     private String measure;
@@ -16,11 +16,11 @@ public class Ingredient {
     @Json(name = "ingredient")
     private String ingredient;
 
-    public Integer getQuantity() {
+    public double getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(Integer quantity) {
+    public void setQuantity(double quantity) {
         this.quantity = quantity;
     }
